Refactor Customize-Snapshot and Customize-Snapshot-Manager for 4.7 support [WIP] [ci skip]
