+++ php/class-customize-snapshot-manager.php
@@ -33,7 +33,7 @@ class Customize_Snapshot_Manager {
 	/**
 	 * Post type.
 	 *
-	 * @var Post_Type
+	 * @var Post_Type|Post_Type_Back_Compt
 	 */
 	public $post_type;
 
@@ -42,7 +42,7 @@ class Customize_Snapshot_Manager {
 	 *
 	 * @todo Rename `Customize_Snapshot` to just `Snapshot`.
 	 *
-	 * @var Customize_Snapshot
+	 * @var Customize_Snapshot|Customize_Snapshot_Back_Compat
 	 */
 	public $snapshot;
 
@@ -89,21 +89,12 @@ class Customize_Snapshot_Manager {
 	}
 
 	/**
-	 * Init.
+	 * Init common hooks.
 	 *
 	 * @global \WP_Customize_Manager $wp_customize
 	 */
-	function init() {
-		// Todo change this 4.7-beta1 to 4.7.
-		if ( version_compare( get_bloginfo( 'version' ), '4.7-beta1', '>=' ) ) {
-			$this->post_type = new Post_Type( $this );
-		} else {
-			$this->post_type = new Post_Type_Back_Compat( $this );
-		}
+	function hooks() {
 		add_action( 'init', array( $this->post_type, 'init' ) );
-
-		add_action( 'template_redirect', array( $this, 'show_theme_switch_error' ) );
-
 		add_action( 'customize_controls_enqueue_scripts', array( $this, 'enqueue_controls_scripts' ) );
 		add_action( 'customize_preview_init', array( $this, 'customize_preview_init' ) );
 		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_frontend_scripts' ) );
@@ -112,22 +103,45 @@ class Customize_Snapshot_Manager {
 		add_action( 'customize_controls_init', array( $this, 'add_snapshot_uuid_to_return_url' ) );
 		add_action( 'customize_controls_print_footer_scripts', array( $this, 'render_templates' ) );
 		add_action( 'customize_save', array( $this, 'check_customize_publish_authorization' ), 10, 0 );
-		add_filter( 'customize_refresh_nonces', array( $this, 'filter_customize_refresh_nonces' ) );
 		add_action( 'admin_bar_menu', array( $this, 'customize_menu' ), 41 );
 		add_action( 'admin_bar_menu', array( $this, 'remove_all_non_snapshot_admin_bar_links' ), 100000 );
 		add_action( 'wp_before_admin_bar_render', array( $this, 'print_admin_bar_styles' ) );
 		add_filter( 'removable_query_args', array( $this, 'filter_removable_query_args' ) );
 
 		add_filter( 'wp_insert_post_data', array( $this, 'prepare_snapshot_post_content_for_publish' ) );
-		add_action( 'customize_save_after', array( $this, 'publish_snapshot_with_customize_save_after' ) );
-		add_action( 'transition_post_status', array( $this, 'save_settings_with_publish_snapshot' ), 10, 3 );
-		add_action( 'wp_ajax_' . self::AJAX_ACTION, array( $this, 'handle_update_snapshot_request' ) );
+	}
 
+	/**
+	 * Init.
+	 */
+	function init() {
+		$this->post_type = new Post_Type( $this );
+		$this->hooks();
+		add_filter( 'customize_save_response', array( $this, 'add_snapshot_var_to_customize_save' ), 10, 2 );
 		if ( $this->read_current_snapshot_uuid() ) {
 			$this->load_snapshot();
-		} elseif ( is_customize_preview() && isset( $_REQUEST['wp_customize_preview_ajax'] ) && 'true' === $_REQUEST['wp_customize_preview_ajax'] ) {
-			add_action( 'wp_loaded', array( $this, 'setup_preview_ajax_requests' ), 12 );
 		}
+	}
+
+	public function load_snapshot() {
+		$this->ensure_customize_manager();
+		$this->snapshot = new Customize_Snapshot( $this );
+	}
+
+	/**
+	 * Add extra changeset variable to publish date.
+	 *
+	 * @param array                 $response          Ajax response.
+	 * @param \WP_Customize_Manager $customize_manager customize manager object.
+	 *
+	 * @return array Response.
+	 */
+	public function add_snapshot_var_to_customize_save( $response, $customize_manager ) {
+		$changeset_post = get_post( $customize_manager->changeset_post_id() );
+		$response['edit_link'] = $this->snapshot->get_edit_link( $changeset_post->ID );
+		$response['publish_date'] = $changeset_post->post_date;
+		$response['title'] = $changeset_post->post_title;
+		return $response;
 	}
 
 	/**
@@ -136,121 +150,22 @@ class Customize_Snapshot_Manager {
 	 * @returns bool Whether a valid snapshot was read.
 	 */
 	public function read_current_snapshot_uuid() {
-		if ( isset( $_REQUEST['customize_snapshot_uuid'] ) ) { // WPCS: input var ok.
-			$uuid = sanitize_key( wp_unslash( $_REQUEST['customize_snapshot_uuid'] ) ); // WPCS: input var ok.
+		$customize_arg = $this->get_customize_uuid_param();
+		$frontend_arg = $this->get_front_uuid_param();
+		if ( isset( $_REQUEST[ $customize_arg ] ) ) {
+			$uuid = $_REQUEST[ $customize_arg ]; // WPCS: input var ok.
+		} elseif ( isset( $_REQUEST[ $frontend_arg ] ) ) {
+			$uuid = $_REQUEST[ $frontend_arg ]; // WPCS: input var ok.
+		}
+
+		if ( isset( $uuid ) ) {
+			$uuid = sanitize_key( wp_unslash( $uuid ) );
 			if ( static::is_valid_uuid( $uuid ) ) {
 				$this->current_snapshot_uuid = $uuid;
 				return true;
 			}
 		}
 		$this->current_snapshot_uuid = null;
-		return false;
-	}
-
-	/**
-	 * Load snapshot.
-	 */
-	public function load_snapshot() {
-		$this->ensure_customize_manager();
-		$this->snapshot = new Customize_Snapshot( $this, $this->current_snapshot_uuid );
-
-		if ( ! $this->should_import_and_preview_snapshot( $this->snapshot ) ) {
-			return;
-		}
-
-		$this->add_widget_setting_preview_filters();
-		$this->add_nav_menu_setting_preview_filters();
-
-		/*
-		 * Populate post values.
-		 *
-		 * Note we have to defer until setup_theme since the transaction
-		 * can be set beforehand, and wp_magic_quotes() would not have
-		 * been called yet, resulting in a $_POST['customized'] that is
-		 * double-escaped. Note that this happens at priority 1, which
-		 * is immediately after Customize_Snapshot_Manager::store_customized_post_data
-		 * which happens at setup_theme priority 0, so that the initial
-		 * POST data can be preserved.
-		 */
-		if ( did_action( 'setup_theme' ) ) {
-			$this->import_snapshot_data();
-		} else {
-			add_action( 'setup_theme', array( $this, 'import_snapshot_data' ) );
-		}
-
-		// Block the robots.
-		add_action( 'wp_head', 'wp_no_robots' );
-
-		// Preview post values.
-		if ( did_action( 'wp_loaded' ) ) {
-			$this->preview_snapshot_settings();
-		} else {
-			add_action( 'wp_loaded', array( $this, 'preview_snapshot_settings' ), 11 );
-		}
-	}
-
-	/**
-	 * Setup previewing of Ajax requests in the Customizer preview.
-	 *
-	 * @global \WP_Customize_Manager $wp_customize
-	 */
-	public function setup_preview_ajax_requests() {
-		global $wp_customize, $pagenow;
-
-		/*
-		 * When making admin-ajax requests from the frontend, settings won't be
-		 * previewed because is_admin() and the call to preview will be
-		 * short-circuited in \WP_Customize_Manager::wp_loaded().
-		 */
-		if ( ! did_action( 'customize_preview_init' ) ) {
-			$wp_customize->customize_preview_init();
-		}
-
-		// Note that using $pagenow is easier to test vs DOING_AJAX.
-		if ( ! empty( $pagenow ) && 'admin-ajax.php' === $pagenow ) {
-			$this->override_request_method();
-		} else {
-			add_action( 'parse_request', array( $this, 'override_request_method' ), 5 );
-		}
-
-		$wp_customize->remove_preview_signature();
-	}
-
-	/**
-	 * Attempt to convert the current request environment into another environment.
-	 *
-	 * @global \WP $wp
-	 *
-	 * @return bool Whether the override was applied.
-	 */
-	public function override_request_method() {
-		global $wp;
-
-		// Skip of X-HTTP-Method-Override request header is not present.
-		if ( ! isset( $_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE'] ) ) {
-			return false;
-		}
-
-		// Skip if REST API request since it has built-in support for overriding the request method.
-		if ( ! empty( $wp ) && ! empty( $wp->query_vars['rest_route'] ) ) {
-			return false;
-		}
-
-		// Skip if the request method is not GET or POST, or the override is the same as the original.
-		$original_request_method = $_SERVER['REQUEST_METHOD'];
-		$override_request_method = strtoupper( $_SERVER['HTTP_X_HTTP_METHOD_OVERRIDE'] );
-		if ( ! in_array( $override_request_method, array( 'GET', 'POST' ), true ) || $original_request_method === $override_request_method ) {
-			return false;
-		}
-
-		// Convert a POST request into a GET request.
-		if ( 'GET' === $override_request_method && 'POST' === $original_request_method ) {
-			$_SERVER['REQUEST_METHOD'] = $override_request_method;
-			$_GET = array_merge( $_GET, $_POST );
-			$_SERVER['QUERY_STRING'] = build_query( array_map( 'rawurlencode', wp_unslash( $_GET ) ) );
-			return true;
-		}
-
 		return false;
 	}
 
@@ -272,7 +187,11 @@ class Customize_Snapshot_Manager {
 		global $wp_customize;
 		if ( empty( $wp_customize ) || ! ( $wp_customize instanceof \WP_Customize_Manager ) ) {
 			require_once( ABSPATH . WPINC . '/class-wp-customize-manager.php' );
-			$wp_customize = new \WP_Customize_Manager(); // WPCS: override ok.
+			if ( null !== $this->current_snapshot_uuid ) {
+				$wp_customize = new \WP_Customize_Manager( array( 'changeset_uuid' => $this->current_snapshot_uuid ) ); // WPCS: override ok.
+			} else {
+				$wp_customize = new \WP_Customize_Manager(); // WPCS: override ok.
+			}
 		}
 
 		$this->customize_manager = $wp_customize;
@@ -282,6 +201,8 @@ class Customize_Snapshot_Manager {
 	 * Is previewing another theme.
 	 *
 	 * @return bool Whether theme is active.
+	 *
+	 * @todo move to back compat?
 	 */
 	public function is_theme_active() {
 		if ( empty( $this->customize_manager ) ) {
@@ -291,267 +212,11 @@ class Customize_Snapshot_Manager {
 	}
 
 	/**
-	 * Determine whether the current snapshot can be previewed.
-	 *
-	 * @param Customize_Snapshot $snapshot Snapshot to check.
-	 * @return true|\WP_Error Returns true if previewable, or `WP_Error` if cannot.
-	 */
-	public function should_import_and_preview_snapshot( Customize_Snapshot $snapshot ) {
-		global $pagenow;
-
-		// Ignore if in the admin, but not Admin Ajax or Customizer.
-		if ( is_admin() && ! in_array( $pagenow, array( 'admin-ajax.php', 'customize.php' ), true ) ) {
-			return false;
-		}
-
-		if ( is_wp_error( $this->get_theme_switch_error( $snapshot ) ) ) {
-			return false;
-		}
-
-		// Abort if doing customize_save.
-		if ( $this->doing_customize_save_ajax() ) {
-			return false;
-		}
-
-		// Abort if the snapshot was already published.
-		if ( $snapshot->saved() && 'publish' === get_post_status( $snapshot->post() ) ) {
-			return false;
-		}
-
-		/*
-		 * Prevent clobbering existing values (or previewing non-snapshotted values on frontend).
-		 * Note that wp.customize.Snapshots.extendPreviewerQuery() will extend the
-		 * previewer data to include the current snapshot UUID.
-		 */
-		if ( $this->customize_manager && count( $this->customize_manager->unsanitized_post_values() ) > 0 ) {
-			return false;
-		}
-
-		return true;
-	}
-
-	/**
-	 * Populate post values and $_POST['customized'] wth the snapshot's data.
-	 *
-	 * Plugins used to have to dynamically register settings by inspecting the
-	 * $_POST['customized'] var and manually re-parse and inspect to see if it
-	 * contains settings that wouldn't be registered otherwise. This ensures
-	 * that these plugins will continue to work.
-	 *
-	 * Note that this can't be called prior to the setup_theme action or else
-	 * magic quotes may end up getting added twice.
-	 *
-	 * @see Customize_Snapshot_Manager::should_import_and_preview_snapshot()
-	 */
-	public function import_snapshot_data() {
-		/*
-		 * We don't merge the snapshot data with any existing existing unsanitized
-		 * post values since should_import_and_preview_snapshot returns false if
-		 * there is any existing data in the Customizer state. This is to prevent
-		 * clobbering existing values (or previewing non-snapshotted values on frontend).
-		 * Note that wp.customize.Snapshots.extendPreviewerQuery() will extend the
-		 * previewer data to include the current snapshot UUID.
-		 */
-		$snapshot_values = array_filter(
-			wp_list_pluck( $this->snapshot->data(), 'value' ),
-			function( $value ) {
-				return ! is_null( $value );
-			}
-		);
-
-		// Populate input vars for back-compat.
-		$_POST['customized'] = wp_slash( wp_json_encode( $snapshot_values ) );
-		// @codingStandardsIgnoreStart
-		$_REQUEST['customized'] = $_POST['customized'];
-		// @codingStandardsIgnoreEnd
-
-		foreach ( $snapshot_values as $setting_id => $value ) {
-			$this->customize_manager->set_post_value( $setting_id, $value );
-		}
-	}
-
-	/**
-	 * Is previewing settings.
-	 *
-	 * Plugins and themes may currently only use `is_customize_preview()` to
-	 * decide whether or not they can store a value in the object cache. For
-	 * example, see `Twenty_Eleven_Ephemera_Widget::widget()`. However, when
-	 * viewing a snapshot on the frontend, the `is_customize_preview()` method
-	 * will return `false`. Plugins and themes that store values in the object
-	 * cache must either skip doing this when `$this->previewing` is `true`,
-	 * or include the `$this->current_snapshot_uuid` (`current_snapshot_uuid()`)
-	 * in the cache key when it is `true`. Note that if the `customize_preview_init` action
-	 * was done, this means that the settings have been previewed in the regular
-	 * Customizer preview.
-	 *
-	 * @see Twenty_Eleven_Ephemera_Widget::widget()
-	 * @see WP_Customize_Manager::is_previewing_settings()
-	 * @see is_previewing_settings()
-	 * @see current_snapshot_uuid()()
-	 * @see WP_Customize_Manager::customize_preview_init()
-	 * @see Customize_Snapshot_Manager::$previewing_settings
-	 *
-	 * @return bool Whether previewing settings.
-	 */
-	public function is_previewing_settings() {
-		return $this->previewing_settings || did_action( 'customize_preview_init' );
-	}
-
-	/**
-	 * Preview the snapshot settings.
-	 *
-	 * Note that this happens at `wp_loaded` action with priority 11 so that we
-	 * can look at whether the `customize_preview_init` action was done.
-	 */
-	public function preview_snapshot_settings() {
-		if ( $this->is_previewing_settings() ) {
-			return;
-		}
-		$this->previewing_settings = true;
-
-		/*
-		 * Note that we need to preview the settings outside the Customizer preview
-		 * and in the Customizer pane itself so we can load a previous snapshot
-		 * into the Customizer. We have to prevent the previews from being added
-		 * in the case of a customize_save action because then update_option()
-		 * may short-circuit because it will detect that there are no changes to
-		 * make.
-		 */
-		foreach ( $this->snapshot->settings() as $setting ) {
-			$setting->preview();
-			$setting->dirty = true;
-		}
-	}
-
-	/**
-	 * Add filters for previewing widgets on the frontend.
-	 */
-	function add_widget_setting_preview_filters() {
-		/*
-		 * Add WP_Customize_Widget component hooks which were short-circuited in 4.5 (r36611 for #35895).
-		 * See https://core.trac.wordpress.org/ticket/35895
-		 */
-		if ( isset( $this->customize_manager->widgets ) && ! current_user_can( 'edit_theme_options' ) ) {
-			$hooks = array(
-				'customize_dynamic_setting_args' => array(
-					'callback' => array( $this->customize_manager->widgets, 'filter_customize_dynamic_setting_args' ),
-					'priority' => 10,
-				),
-				'widgets_init' => array(
-					'callback' => array( $this->customize_manager->widgets, 'register_settings' ),
-					'priority' => 95,
-				),
-				'customize_register' => array(
-					'callback' => array( $this->customize_manager->widgets, 'schedule_customize_register' ),
-					'priority' => 1,
-				),
-			);
-			foreach ( $hooks as $hook_name => $hook_args ) {
-				// Note that add_action()/has_action() are just aliases for add_filter()/has_filter().
-				if ( ! has_filter( $hook_name, $hook_args['callback'] ) ) {
-					add_filter( $hook_name, $hook_args['callback'], $hook_args['priority'], PHP_INT_MAX );
-				}
-			}
-		}
-
-		/*
-		 * Disable routine which fails because \WP_Customize_Manager::setup_theme() is
-		 * never called in a frontend preview context, whereby the original_stylesheet
-		 * is never set and so \WP_Customize_Manager::is_theme_active() will thus
-		 * always return true because get_stylesheet() !== null.
-		 *
-		 * The action being removed is responsible for adding an option_sidebar_widgets
-		 * filter \WP_Customize_Widgets::filter_option_sidebars_widgets_for_theme_switch()
-		 * which causes the sidebars_widgets to be overridden with a global variable.
-		 */
-		if ( ! is_admin() ) {
-			remove_action( 'wp_loaded', array( $this->customize_manager->widgets, 'override_sidebars_widgets_for_theme_switch' ) );
-		}
-	}
-
-	/**
-	 * Add filters for previewing nav menus on the frontend.
-	 */
-	public function add_nav_menu_setting_preview_filters() {
-		if ( isset( $this->customize_manager->nav_menus ) && ! current_user_can( 'edit_theme_options' ) ) {
-			$hooks = array(
-				'customize_register' => array(
-					'callback' => array( $this->customize_manager->nav_menus, 'customize_register' ),
-					'priority' => 11,
-				),
-				'customize_dynamic_setting_args' => array(
-					'callback' => array( $this->customize_manager->nav_menus, 'filter_dynamic_setting_args' ),
-					'priority' => 10,
-				),
-				'customize_dynamic_setting_class' => array(
-					'callback' => array( $this->customize_manager->nav_menus, 'filter_dynamic_setting_class' ),
-					'priority' => 10,
-				),
-				'wp_nav_menu_args' => array(
-					'callback' => array( $this->customize_manager->nav_menus, 'filter_wp_nav_menu_args' ),
-					'priority' => 1000,
-				),
-				'wp_nav_menu' => array(
-					'callback' => array( $this->customize_manager->nav_menus, 'filter_wp_nav_menu' ),
-					'priority' => 10,
-				),
-			);
-			foreach ( $hooks as $hook_name => $hook_args ) {
-				// Note that add_action()/has_action() are just aliases for add_filter()/has_filter().
-				if ( ! has_filter( $hook_name, $hook_args['callback'] ) ) {
-					add_filter( $hook_name, $hook_args['callback'], $hook_args['priority'], PHP_INT_MAX );
-				}
-			}
-		}
-
-		if ( isset( $this->customize_manager->nav_menus ) ) {
-			add_action( 'customize_register', array( $this, 'preview_early_nav_menus_in_customizer' ), 9 );
-		}
-	}
-
-	/**
-	 * Preview nav menu settings early so that the sections and controls for snapshot values will be added properly.
-	 *
-	 * This must happen at `customize_register` priority prior to 11 which is when `WP_Customize_Nav_Menus::customize_register()` runs.
-	 * This is only relevant when accessing the Customizer app (customize.php), as this is where sections/controls matter.
-	 *
-	 * @see \WP_Customize_Nav_Menus::customize_register()
-	 */
-	public function preview_early_nav_menus_in_customizer() {
-		if ( ! is_admin() ) {
-			return;
-		}
-		$this->customize_manager->add_dynamic_settings( array_keys( $this->snapshot()->data() ) );
-		foreach ( $this->snapshot->settings() as $setting ) {
-			$is_nav_menu_setting = (
-				$setting instanceof \WP_Customize_Nav_Menu_Setting
-				||
-				$setting instanceof \WP_Customize_Nav_Menu_Item_Setting
-				||
-				preg_match( '/^nav_menu_locations\[/', $setting->id )
-			);
-			if ( $is_nav_menu_setting ) {
-				$setting->preview();
-
-				/*
-				 * The following is redundant because it will be done later in
-				 * Customize_Snapshot_Manager::preview_snapshot_settings().
-				 * Also note that the $setting instance here will likely be
-				 * blown away inside of WP_Customize_Nav_Menus::customize_register(),
-				 * when add_setting is called there. What matters here is that
-				 * preview() is called on the setting _before_ the logic inside
-				 * WP_Customize_Nav_Menus::customize_register() runs, so that
-				 * the nav menu sections will be created.
-				 */
-				$setting->dirty = true;
-			}
-		}
-	}
-
-	/**
 	 * Add snapshot UUID the Customizer return URL.
 	 *
 	 * If the Customizer was loaded with a snapshot UUID, let the return URL include this snapshot.
+	 *
+	 * @todo move to back compat?
 	 */
 	public function add_snapshot_uuid_to_return_url() {
 		$should_add_snapshot_uuid = (
@@ -562,8 +227,9 @@ class Customize_Snapshot_Manager {
 			false === strpos( $this->customize_manager->get_return_url(), '/wp-admin/' )
 		);
 		if ( $should_add_snapshot_uuid ) {
+			$args_name = $this->get_front_uuid_param();
 			$args = array(
-				'customize_snapshot_uuid' => $this->current_snapshot_uuid,
+				$args_name => $this->current_snapshot_uuid,
 			);
 			$return_url = add_query_arg( array_map( 'rawurlencode', $args ), $this->customize_manager->get_return_url() );
 			$this->customize_manager->set_return_url( $return_url );
@@ -571,46 +237,9 @@ class Customize_Snapshot_Manager {
 	}
 
 	/**
-	 * Show the theme switch error if there is one.
-	 */
-	public function show_theme_switch_error() {
-		if ( empty( $this->snapshot ) ) {
-			return;
-		}
-		$error = $this->get_theme_switch_error( $this->snapshot );
-		if ( is_wp_error( $error ) ) {
-			wp_die( esc_html( $error->get_error_message() ) );
-		}
-	}
-
-	/**
-	 * Redirect when preview is not allowed for the current theme.
-	 *
-	 * @param Customize_Snapshot $snapshot Snapshot to check.
-	 * @return \WP_Error|null
-	 */
-	public function get_theme_switch_error( Customize_Snapshot $snapshot ) {
-
-		// Loading a snapshot into the context of a theme switch is not supported.
-		if ( ! $this->is_theme_active() ) {
-			return new \WP_Error( 'snapshot_theme_switch', __( 'Snapshot cannot be previewed when previewing a theme switch.', 'customize-snapshots' ) );
-		}
-
-		$snapshot_post = $snapshot->post();
-		if ( ! $snapshot_post ) {
-			return null;
-		}
-
-		$snapshot_theme = get_post_meta( $snapshot_post->ID, '_snapshot_theme', true );
-		if ( ! empty( $snapshot_theme ) && get_stylesheet() !== $snapshot_theme ) {
-			return new \WP_Error( 'snapshot_theme_switched', __( 'Snapshot requested was made for a different theme and cannot be previewed with the current theme.', 'customize-snapshots' ) );
-		}
-
-		return null;
-	}
-
-	/**
 	 * Check whether customize_publish capability is granted in customize_save.
+	 *
+	 * @todo move to back compat?
 	 */
 	public function check_customize_publish_authorization() {
 		if ( $this->doing_customize_save_ajax() && ! current_user_can( 'customize_publish' ) ) {
@@ -628,9 +257,9 @@ class Customize_Snapshot_Manager {
 	 */
 	static public function encode_json( $value ) {
 		$flags = 0;
-		if ( defined( '\JSON_PRETTY_PRINT' ) ) {
-			$flags |= \JSON_PRETTY_PRINT;
-		}
+
+		$flags |= \JSON_PRETTY_PRINT;
+
 		if ( defined( '\JSON_UNESCAPED_SLASHES' ) ) {
 			$flags |= \JSON_UNESCAPED_SLASHES;
 		}
@@ -645,23 +274,18 @@ class Customize_Snapshot_Manager {
 	 */
 	public function enqueue_controls_scripts() {
 
-		// Prevent loading the Snapshot interface if the theme is not active.
-		if ( ! $this->is_theme_active() ) {
-			return;
-		}
-
 		wp_enqueue_style( 'customize-snapshots' );
 		wp_enqueue_script( 'customize-snapshots' );
 		if ( $this->snapshot ) {
-			$post = $this->snapshot->post();
+			$post_id = $this->customize_manager->changeset_post_id();
+			$post = get_post( $post_id );
 			$this->override_post_date_default_data( $post );
+			$edit_link = $this->snapshot->get_edit_link( $post );
 		}
 
 		// Script data array.
 		$exports = apply_filters( 'customize_snapshots_export_data', array(
-			'action' => self::AJAX_ACTION,
-			'uuid' => $this->snapshot ? $this->snapshot->uuid() : self::generate_uuid(),
-			'editLink' => isset( $post ) ? get_edit_post_link( $post, 'raw' ) : '',
+			'editLink' => isset( $edit_link ) ? $edit_link : '',
 			'publishDate' => isset( $post->post_date ) ? $post->post_date : '',
 			'title' => isset( $post->post_title ) ? $post->post_title : '',
 			'postStatus' => isset( $post->post_status ) ? $post->post_status : '',
@@ -685,7 +309,6 @@ class Customize_Snapshot_Manager {
 				'collapseSnapshotScheduling' => __( 'Collapse snapshot scheduling', 'customize-snapshots' ),
 				'expandSnapshotScheduling' => __( 'Expand snapshot scheduling', 'customize-snapshots' ),
 			),
-			'snapshotExists' => ( $this->snapshot && $this->snapshot->saved() ),
 		) );
 
 		// Export data to JS.
@@ -778,95 +401,12 @@ class Customize_Snapshot_Manager {
 	}
 
 	/**
-	 * Include the snapshot nonce in the Customizer nonces.
-	 *
-	 * @param array $nonces Nonces.
-	 * @return array Nonces.
-	 */
-	public function filter_customize_refresh_nonces( $nonces ) {
-		$nonces['snapshot'] = wp_create_nonce( self::AJAX_ACTION );
-		return $nonces;
-	}
-
-	/**
 	 * Get the Customize_Snapshot instance.
 	 *
 	 * @return Customize_Snapshot
 	 */
 	public function snapshot() {
 		return $this->snapshot;
-	}
-
-	/**
-	 * Publish the snapshot snapshots via AJAX.
-	 *
-	 * Fires at `customize_save_after` to update and publish the snapshot.
-	 * The logic in here is the inverse of save_settings_with_publish_snapshot.
-	 *
-	 * @see Customize_Snapshot_Manager::save_settings_with_publish_snapshot()
-	 *
-	 * @return bool Whether the snapshot was saved successfully.
-	 */
-	public function publish_snapshot_with_customize_save_after() {
-		$that = $this;
-
-		if ( ! $this->snapshot || ! $this->doing_customize_save_ajax() ) {
-			return false;
-		}
-
-		// This should never be reached due to Customize_Snapshot_Manager::check_customize_publish_authorization().
-		if ( ! current_user_can( 'customize_publish' ) ) {
-			return false;
-		}
-
-		$settings_data = array_map(
-			function( $value ) {
-				return compact( 'value' );
-			},
-			$this->customize_manager->unsanitized_post_values()
-		);
-		$result = $this->snapshot->set( $settings_data, array( 'skip_validation' => true ) );
-		if ( ! empty( $result['errors'] ) ) {
-			add_filter( 'customize_save_response', function( $response ) use ( $result, $that ) {
-				$response['snapshot_errors'] = $that->prepare_errors_for_response( $result['errors'] );
-				return $response;
-			} );
-			return false;
-		}
-
-		if ( ! $this->snapshot->post() || 'publish' !== $this->snapshot->post()->post_status ) {
-			$args = array(
-				'status' => 'publish',
-			);
-
-			// Ensure a scheduled Snapshot is published.
-			if ( $this->snapshot->post() && 'future' === $this->snapshot->post()->post_status ) {
-				$args['edit_date'] = true;
-				$args['post_date'] = current_time( 'mysql', false );
-				$args['post_date_gmt'] = current_time( 'mysql', true );
-			}
-
-			if ( isset( $_POST['title'] ) && '' !== trim( $_POST['title'] ) ) {
-				$args['post_title'] = sanitize_text_field( wp_unslash( $_POST['title'] ) );
-			}
-
-			$r = $this->snapshot->save( $args );
-			if ( is_wp_error( $r ) ) {
-				add_filter( 'customize_save_response', function( $response ) use ( $r, $that ) {
-					$response['snapshot_errors'] = $that->prepare_errors_for_response( $r );
-					return $response;
-				} );
-				return false;
-			}
-		}
-
-		// Send the new UUID to the client for the next snapshot.
-		$class = __CLASS__; // For PHP 5.3.
-		add_filter( 'customize_save_response', function( $data ) use ( $class ) {
-			$data['new_customize_snapshot_uuid'] = $class::generate_uuid();
-			return $data;
-		} );
-		return true;
 	}
 
 	/**
@@ -927,314 +467,6 @@ class Customize_Snapshot_Manager {
 	}
 
 	/**
-	 * Publish snapshot changes when snapshot post is being published.
-	 *
-	 * The logic in here is the inverse of to publish_snapshot_with_customize_save_after.
-	 *
-	 * The meat of the logic that manipulates the post_content and validates the settings
-	 * needs to be done in wp_insert_post_data filter in like a
-	 * filter_insert_post_data_to_validate_published_snapshot method? This would
-	 * have the benefit of reducing one wp_insert_post() call.
-	 *
-	 * @todo Consider using wp_insert_post_data to prevent double calls to wp_insert_post().
-	 * @see Customize_Snapshot_Manager::publish_snapshot_with_customize_save_after()
-	 *
-	 * @param string   $new_status New status.
-	 * @param string   $old_status Old status.
-	 * @param \WP_Post $post       Post object.
-	 * @return bool Whether the settings were saved.
-	 */
-	public function save_settings_with_publish_snapshot( $new_status, $old_status, $post ) {
-
-		// Abort if not transitioning a snapshot post to publish from a non-publish status.
-		if ( $this->get_post_type() !== $post->post_type || 'publish' !== $new_status || $new_status === $old_status ) {
-			return false;
-		}
-
-		$this->ensure_customize_manager();
-
-		if ( $this->doing_customize_save_ajax() ) {
-			// Short circuit because customize_save ajax call is changing status.
-			return false;
-		}
-
-		if ( ! did_action( 'customize_register' ) ) {
-			/*
-			 * When running from CLI or Cron, we have to remove the action because
-			 * it will get added with a default priority of 10, after themes and plugins
-			 * have already done add_action( 'customize_register' ), resulting in them
-			 * being called first at the priority 10. So we manually call the
-			 * prerequisite function WP_Customize_Manager::register_controls() and
-			 * remove it from being called when the customize_register action fires.
-			 */
-			remove_action( 'customize_register', array( $this->customize_manager, 'register_controls' ) );
-			$this->customize_manager->register_controls();
-
-			/*
-			 * Unfortunate hack to prevent \WP_Customize_Widgets::customize_register()
-			 * from calling preview() on settings. This needs to be cleaned up in core.
-			 * It is important for previewing to be prevented because if an option has
-			 * a filter it will short-circuit when an update is attempted since it
-			 * detects that there is no change to be put into the DB.
-			 * See: https://github.com/xwp/wordpress-develop/blob/e8c58c47db1421a1d0b2afa9ad4b9eb9e1e338e0/src/wp-includes/class-wp-customize-widgets.php#L208-L217
-			 */
-			if ( ! defined( 'DOING_AJAX' ) ) {
-				define( 'DOING_AJAX', true );
-			}
-			$_REQUEST['action'] = 'customize_save';
-
-			/** This action is documented in wp-includes/class-wp-customize-manager.php */
-			do_action( 'customize_register', $this->customize_manager );
-
-			// undefine( 'DOING_AJAX' )... just kidding. This is the end of the unfortunate hack and it should be fixed in Core.
-			unset( $_REQUEST['action'] );
-		}
-		$snapshot_content = $this->post_type->get_post_content( $post );
-
-		if ( method_exists( $this->customize_manager, 'validate_setting_values' ) ) {
-			/** This action is documented in wp-includes/class-wp-customize-manager.php */
-			do_action( 'customize_save_validation_before', $this->customize_manager );
-		}
-
-		$setting_ids = array_keys( $snapshot_content );
-		$this->customize_manager->add_dynamic_settings( $setting_ids );
-
-		/** This action is documented in wp-includes/class-wp-customize-manager.php */
-		do_action( 'customize_save', $this->customize_manager );
-
-		/**
-		 * Settings to save.
-		 *
-		 * @var \WP_Customize_Setting[]
-		 */
-		$settings = array();
-
-		$publish_error_count = 0;
-		foreach ( $snapshot_content as $setting_id => &$setting_params ) {
-
-			// Missing value error.
-			if ( ! isset( $setting_params['value'] ) || is_null( $setting_params['value'] ) ) {
-				if ( ! is_array( $setting_params ) ) {
-					if ( ! empty( $setting_params ) ) {
-						$setting_params = array( 'value' => $setting_params );
-					} else {
-						$setting_params = array();
-					}
-				}
-				$setting_params['publish_error'] = 'null_value';
-				$publish_error_count += 1;
-				continue;
-			}
-
-			// Unrecognized setting error.
-			$this->customize_manager->set_post_value( $setting_id, $setting_params['value'] );
-			$setting = $this->customize_manager->get_setting( $setting_id );
-			if ( ! ( $setting instanceof \WP_Customize_Setting ) ) {
-				$setting_params['publish_error'] = 'unrecognized_setting';
-				$publish_error_count += 1;
-				continue;
-			}
-
-			// Validate setting value.
-			if ( method_exists( $setting, 'validate' ) ) {
-				$validity = $setting->validate( $setting_params['value'] );
-				if ( is_wp_error( $validity ) ) {
-					$setting_params['publish_error'] = $validity->get_error_code();
-					$publish_error_count += 1;
-					continue;
-				}
-			}
-
-			// Validate sanitized setting value.
-			$sanitized_value = $setting->sanitize( $setting_params['value'] );
-			if ( is_null( $sanitized_value ) || is_wp_error( $sanitized_value ) ) {
-				$setting_params['publish_error'] = is_wp_error( $sanitized_value ) ? $sanitized_value->get_error_code() : 'invalid_value';
-				$publish_error_count += 1;
-				continue;
-			}
-
-			$settings[] = $setting;
-			unset( $setting_params['publish_error'] );
-		}
-
-		// Handle error scenarios.
-		if ( $publish_error_count > 0 ) {
-			$update_setting_args = array(
-				'ID' => $post->ID,
-				'post_content' => Customize_Snapshot_Manager::encode_json( $snapshot_content ),
-				'post_status' => 'pending',
-			);
-			wp_update_post( wp_slash( $update_setting_args ) );
-			update_post_meta( $post->ID, 'snapshot_error_on_publish', $publish_error_count );
-
-			add_filter( 'redirect_post_location', function( $location ) {
-				$location = add_query_arg( 'snapshot_error_on_publish', '1', $location );
-				return $location;
-			} );
-			return false;
-		}
-
-		/*
-		 * Change all setting capabilities temporarily to 'exist' to allow them to
-		 * be saved regardless of current user, such as when WP-Cron is publishing
-		 * the snapshot post if it was scheduled. It is safe to do this because
-		 * a setting can only be written into a snapshot by users who have the
-		 * capability, so after it has been added to a snapshot it is good to commit.
-		 */
-		$existing_caps = wp_list_pluck( $settings, 'capability' );
-		foreach ( $settings as $setting ) {
-			$setting->capability = 'exist';
-		}
-
-		// Persist the settings in the DB.
-		foreach ( $settings as $setting ) {
-			$setting->save();
-		}
-
-		// Restore setting capabilities.
-		foreach ( $existing_caps as $setting_id => $existing_cap ) {
-			$settings[ $setting_id ]->capability = $existing_cap;
-		}
-
-		/** This action is documented in wp-includes/class-wp-customize-manager.php */
-		do_action( 'customize_save_after', $this->customize_manager );
-
-		// Remove any previous error on setting.
-		delete_post_meta( $post->ID, 'snapshot_error_on_publish' );
-
-		return true;
-	}
-
-	/**
-	 * Update snapshots via AJAX.
-	 */
-	public function handle_update_snapshot_request() {
-		if ( ! check_ajax_referer( self::AJAX_ACTION, 'nonce', false ) ) {
-			status_header( 400 );
-			wp_send_json_error( 'bad_nonce' );
-		} elseif ( ! current_user_can( 'customize' ) ) {
-			status_header( 403 );
-			wp_send_json_error( 'customize_not_allowed' );
-		} elseif ( ! isset( $_SERVER['REQUEST_METHOD'] ) || 'POST' !== $_SERVER['REQUEST_METHOD'] ) { // WPCS: input var ok.
-			status_header( 405 );
-			wp_send_json_error( 'bad_method' );
-		} elseif ( empty( $this->current_snapshot_uuid ) ) {
-			status_header( 400 );
-			wp_send_json_error( 'invalid_customize_snapshot_uuid' );
-		} elseif ( 0 === count( $this->customize_manager->unsanitized_post_values() ) ) {
-			status_header( 400 );
-			wp_send_json_error( 'missing_snapshot_customized' );
-		}
-
-		if ( isset( $_POST['status'] ) ) { // WPCS: input var ok.
-			$status = sanitize_key( $_POST['status'] );
-		} else {
-			$status = 'draft';
-		}
-		if ( ! in_array( $status, array( 'draft', 'pending', 'future' ), true ) ) {
-			status_header( 400 );
-			wp_send_json_error( 'bad_status' );
-		}
-		if ( 'future' === $status && ! current_user_can( 'customize_publish' ) ) {
-			status_header( 400 );
-			wp_send_json_error( 'customize_not_allowed' );
-		}
-		$publish_date = isset( $_POST['publish_date'] ) ? $_POST['publish_date'] : '';
-		if ( 'future' === $status ) {
-			$publish_date_obj = new \DateTime( $publish_date );
-			$current_date = new \DateTime( current_time( 'mysql' ) );
-			if ( empty( $publish_date ) || ! $publish_date_obj || $publish_date > $current_date ) {
-				status_header( 400 );
-				wp_send_json_error( 'bad_schedule_time' );
-			}
-		}
-
-		// Prevent attempting to modify a "locked" snapshot (a published one).
-		$post = $this->snapshot->post();
-		if ( $post && 'publish' === $post->post_status ) {
-			wp_send_json_error( array(
-				'errors' => array(
-					'already_published' => array(
-						'message' => __( 'The snapshot has already published so it is locked.', 'customize-snapshots' ),
-					),
-				),
-			) );
-		}
-
-		/**
-		 * Add any additional checks before saving snapshot.
-		 *
-		 * @param Customize_Snapshot         $snapshot         Snapshot to be saved.
-		 * @param Customize_Snapshot_Manager $snapshot_manager Snapshot manager.
-		 */
-		do_action( 'customize_snapshot_save_before', $this->snapshot, $this );
-
-		// Set the snapshot UUID.
-		$post_type = get_post_type_object( $this->get_post_type() );
-		$authorized = ( $post ?
-			current_user_can( $post_type->cap->edit_post, $post->ID ) :
-			current_user_can( 'customize' )
-		);
-		if ( ! $authorized ) {
-			status_header( 403 );
-			wp_send_json_error( 'unauthorized' );
-		}
-
-		$data = array(
-			'errors' => null,
-		);
-		$settings_data = array_map(
-			function( $value ) {
-				return compact( 'value' );
-			},
-			$this->customize_manager->unsanitized_post_values()
-		);
-		$r = $this->snapshot->set( $settings_data );
-		if ( method_exists( $this->customize_manager, 'prepare_setting_validity_for_js' ) ) {
-			$data['setting_validities'] = array_map(
-				array( $this->customize_manager, 'prepare_setting_validity_for_js' ),
-				$r['validities']
-			);
-		}
-
-		if ( $r['errors'] ) {
-			$data['errors'] = $this->prepare_errors_for_response( $r['errors'] );
-			wp_send_json_error( $data );
-		}
-		$args = array(
-			'status' => $status,
-		);
-		if ( isset( $_POST['title'] ) && '' !== trim( $_POST['title'] ) ) {
-			$args['post_title'] = sanitize_text_field( wp_unslash( $_POST['title'] ) );
-		}
-
-		$args['edit_date'] = current_time( 'mysql' );
-		if ( isset( $publish_date_obj ) && 'future' === $status ) {
-			$args['post_date'] = $publish_date_obj->format( 'Y-m-d H:i:s' );
-			$args['post_date_gmt'] = '0000-00-00 00:00:00';
-		} else {
-			$args['post_date_gmt'] = $args['post_date'] = '0000-00-00 00:00:00';
-		}
-		$r = $this->snapshot->save( $args );
-
-		$post = $this->snapshot->post();
-		if ( $post ) {
-			$data['edit_link'] = get_edit_post_link( $post, 'raw' );
-			$data['snapshot_publish_date'] = $post->post_date;
-			$data['title'] = $post->post_title;
-		}
-
-		if ( is_wp_error( $r ) ) {
-			$data['errors'] = $this->prepare_errors_for_response( $r );
-			wp_send_json_error( $data );
-		}
-
-		/** This filter is documented in wp-includes/class-wp-customize-manager.php */
-		$data = apply_filters( 'customize_save_response', $data, $this->customize_manager );
-		wp_send_json_success( $data );
-	}
-
-	/**
 	 * Prepare a WP_Error for sending to JS.
 	 *
 	 * @param \WP_Error $error Error.
@@ -1284,7 +516,6 @@ class Customize_Snapshot_Manager {
 	 */
 	public function customize_menu( $wp_admin_bar ) {
 		add_action( 'wp_before_admin_bar_render', 'wp_customize_support_script' );
-		$this->replace_customize_link( $wp_admin_bar );
 		$this->add_resume_snapshot_link( $wp_admin_bar );
 		$this->add_post_edit_screen_link( $wp_admin_bar );
 		$this->add_snapshot_exit_link( $wp_admin_bar );
@@ -1313,46 +544,6 @@ class Customize_Snapshot_Manager {
 			}
 		</style>
 		<?php
-	}
-
-	/**
-	 * Replaces the "Customize" link in the Toolbar.
-	 *
-	 * @param \WP_Admin_Bar $wp_admin_bar WP_Admin_Bar instance.
-	 */
-	public function replace_customize_link( $wp_admin_bar ) {
-		if ( empty( $this->snapshot ) ) {
-			return;
-		}
-
-		$customize_node = $wp_admin_bar->get_node( 'customize' );
-		if ( empty( $customize_node ) ) {
-			return;
-		}
-		$customize_changeset_param_name = constant( get_class( $this->post_type ) . '::CUSTOMIZE_UUID_PARAM_NAME' );
-		$frontend_changeset_param_name = constant( get_class( $this->post_type ) . '::FRONT_UUID_PARAM_NAME' );
-
-		// Remove customize_snapshot_uuid query param from url param to be previewed in Customizer.
-		$preview_url_query_params = array();
-		$preview_url_parsed = wp_parse_url( $customize_node->href );
-		parse_str( $preview_url_parsed['query'], $preview_url_query_params );
-		if ( ! empty( $preview_url_query_params['url'] ) ) {
-			$preview_url_query_params['url'] = remove_query_arg( array( $frontend_changeset_param_name ), $preview_url_query_params['url'] );
-			$customize_node->href = preg_replace(
-				'/(?<=\?).*?(?=#|$)/',
-				build_query( $preview_url_query_params ),
-				$customize_node->href
-			);
-		}
-
-		// Add customize_snapshot_uuid param as param to customize.php itself.
-		$customize_node->href = add_query_arg(
-			array( $customize_changeset_param_name => $this->current_snapshot_uuid ),
-			$customize_node->href
-		);
-
-		$customize_node->meta['class'] .= ' ab-customize-snapshots-item';
-		$wp_admin_bar->add_menu( (array) $customize_node );
 	}
 
 	/**
@@ -1387,7 +578,7 @@ class Customize_Snapshot_Manager {
 		$wp_admin_bar->add_menu( array(
 			'id' => 'inspect-customize-snapshot',
 			'title' => __( 'Inspect Snapshot', 'customize-snapshots' ),
-			'href' => get_edit_post_link( $post->ID, 'raw' ),
+			'href' => $this->snapshot->get_edit_link( $post ),
 			'meta' => array(
 				'class' => 'ab-item ab-customize-snapshots-item',
 			),
@@ -1403,12 +594,10 @@ class Customize_Snapshot_Manager {
 		if ( ! $this->snapshot ) {
 			return;
 		}
-		$frontend_changeset_param_name = constant( get_class( $this->post_type ) . '::FRONT_UUID_PARAM_NAME' );
-
 		$wp_admin_bar->add_menu( array(
 			'id' => 'exit-customize-snapshot',
 			'title' => __( 'Exit Snapshot Preview', 'customize-snapshots' ),
-			'href' => remove_query_arg( $frontend_changeset_param_name ),
+			'href' => remove_query_arg( $this->get_front_uuid_param() ),
 			'meta' => array(
 				'class' => 'ab-item ab-customize-snapshots-item',
 			),
@@ -1599,7 +788,6 @@ class Customize_Snapshot_Manager {
 		<?php
 	}
 
-
 	/**
 	 * Format GMT Offset.
 	 *
@@ -1672,8 +860,28 @@ class Customize_Snapshot_Manager {
 
 	/**
 	 * Get Post_Type from dynamic class.
+	 *
+	 * @return string Post type.
 	 */
 	public function get_post_type() {
 		return constant( get_class( $this->post_type ) . '::SLUG' );
+	}
+
+	/**
+	 * Get Frontend UUID param.
+	 *
+	 * @return string param.
+	 */
+	public function get_front_uuid_param() {
+		return constant( get_class( $this->post_type ) . '::FRONT_UUID_PARAM_NAME' );
+	}
+
+	/**
+	 * Get customize uuid param name.
+	 *
+	 * @return string customize param name.
+	 */
+	public function get_customize_uuid_param() {
+		return constant( get_class( $this->post_type ) . '::CUSTOMIZE_UUID_PARAM_NAME' );
 	}
 }
