+++ php/class-post-type.php
@@ -95,6 +95,7 @@ class Post_Type {
 		add_filter( 'map_meta_cap', array( $this, 'remap_customize_meta_cap' ), 5, 4 );
 		add_filter( 'bulk_actions-edit-' . static::SLUG, array( $this, 'add_snapshot_bulk_actions' ) );
 		add_filter( 'handle_bulk_actions-edit-' . static::SLUG, array( $this, 'handle_snapshot_bulk_actions' ), 10, 3 );
+		add_action( 'admin_print_styles-edit.php', array( $this, 'hide_add_new_changeset_button' ) );
 	}
 
 
@@ -670,6 +671,22 @@ class Post_Type {
 			}
 		</style>
 		<?php
+	}
+
+	/**
+	 * Hide add new button for customize_changeset post_type.
+	 */
+	public function hide_add_new_changeset_button() {
+		global $typenow;
+		if ( static::SLUG === $typenow ) {
+			?>
+			<style>
+				a.page-title-action {
+					display: none;
+				}
+			</style>
+			<?php
+		}
 	}
 
 	/**
